# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/workspace/commands": {
      "post": {
        "summary": "Workspace Command",
        "operationId": "workspace_command_api_workspace_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/import": {
      "post": {
        "summary": "Import Local Workspace",
        "operationId": "import_local_workspace_api_workspace_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace/import/{import_id}": {
      "get": {
        "summary": "Import Status",
        "operationId": "import_status_api_workspace_import__import_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Command": {
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "createContext",
              "addResumeVersion",
              "saveMaterial",
              "attachMaterial",
              "detachMaterial",
              "deleteMaterial",
              "deleteContext",
              "saveExperience",
              "saveExperienceDocument",
              "deleteExperienceDocument",
              "setSuggestionFeedback",
              "saveReviewFeedback",
              "setContextExperiences",
              "prepareReview",
              "prepareRetry",
              "restoreRecord",
              "clearActiveRun",
              "registerRun",
              "completeReview"
            ],
            "title": "Operation"
          },
          "args": {
            "items": {},
            "type": "array",
            "maxItems": 4,
            "title": "Args"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation",
          "args"
        ],
        "title": "Command"
      },
...
      "Context": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Name"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "latestVersionId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Latestversionid"
          },
          "latestReviewId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestreviewid"
          },
          "lastAttemptId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastattemptid"
          },
          "selectedExperienceIds": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Selectedexperienceids"
          },
          "userFeedback": {
            "type": "string",
            "maxLength": 2000,
            "title": "Userfeedback"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "createdAt",
          "latestVersionId",
          "userFeedback"
        ],
        "title": "Context"
      },
...
      "ContextMaterial": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 241,
            "minLength": 1,
            "title": "Id"
          },
          "contextId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Contextid"
          },
          "materialId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Materialid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "contextId",
          "materialId"
        ],
        "title": "ContextMaterial"
      },
      "ContextRefs": {
        "properties": {
          "contextId": {
            "type": "string",
            "minLength": 1,
            "title": "Contextid"
          },
          "resumeVersionId": {
            "type": "string",
            "minLength": 1,
            "title": "Resumeversionid"
          },
          "materialVersions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Materialversions"
          },
          "previousReviewId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previousreviewid"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/FeedbackReference"
            },
            "type": "array",
            "title": "Feedback"
          },
          "resolved": {
            "items": {
              "$ref": "#/components/schemas/ResolvedMemory"
            },
            "type": "array",
            "title": "Resolved"
          },
          "userFeedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Userfeedback"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceCandidate"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Experiences"
          },
          "selectedExperienceIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Selectedexperienceids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contextId",
          "resumeVersionId"
        ],
        "title": "ContextRefs"
      },
...
      "Experience": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "period": {
            "type": "string",
            "maxLength": 200,
            "title": "Period"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ExperienceSource"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Sources"
          },
          "markdown": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60000
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12000
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "period",
          "sources",
          "id",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "Experience"
      },
...
      "FileReference": {
        "properties": {
          "fileId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Fileid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fileId"
        ],
        "title": "FileReference"
      },
...
      "FlowDocument": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/FlowBlock"
            },
            "type": "array",
            "minItems": 1,
            "title": "Blocks"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "blocks"
        ],
        "title": "FlowDocument"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRequest": {
        "properties": {
          "importId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Importid"
          },
          "workspace": {
            "$ref": "#/components/schemas/ImportedWorkspace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "importId",
          "workspace"
        ],
        "title": "ImportRequest"
      },
      "ImportedWorkspace": {
        "properties": {
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/Context"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Contexts"
          },
          "resumeVersions": {
            "items": {
              "$ref": "#/components/schemas/ResumeVersion"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Resumeversions"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/Material"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Materials"
          },
          "materialVersions": {
            "items": {
              "$ref": "#/components/schemas/MaterialVersion"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Materialversions"
          },
          "contextMaterials": {
            "items": {
              "$ref": "#/components/schemas/ContextMaterial"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Contextmaterials"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Reviews"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/SuggestionEntry"
            },
            "type": "array",
            "maxItems": 50000,
            "title": "Suggestions"
          },
          "reviewMaterials": {
            "items": {
              "$ref": "#/components/schemas/ReviewMaterial"
            },
            "type": "array",
            "maxItems": 50000,
            "title": "Reviewmaterials"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/Experience"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Experiences"
          },
          "experienceDocuments": {
            "items": {
              "$ref": "#/components/schemas/SavedExperienceDocument"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Experiencedocuments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contexts",
          "resumeVersions",
          "materials",
          "materialVersions",
          "contextMaterials",
          "reviews",
          "suggestions",
          "reviewMaterials",
          "experiences",
          "experienceDocuments"
        ],
        "title": "ImportedWorkspace"
      },
...
      "Material": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "materialType": {
            "type": "string",
            "enum": [
              "company",
              "jobPosting"
            ],
            "title": "Materialtype"
          },
          "currentVersionId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentversionid"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MaterialDraft"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "materialType",
          "currentVersionId",
          "deleted"
        ],
        "title": "Material"
      },
      "MaterialDraft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "materialType": {
            "type": "string",
            "enum": [
              "company",
              "jobPosting"
            ],
            "title": "Materialtype"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "title": "Content"
          },
          "source": {
            "type": "string",
            "maxLength": 2000,
            "title": "Source"
          },
          "original": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileReference"
              },
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Original"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "materialType",
          "content",
          "source"
        ],
        "title": "MaterialDraft"
      },
...
      "MaterialReviewDTO": {
        "properties": {
          "sourceId": {
            "type": "string",
            "minLength": 1,
            "title": "Sourceid"
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          },
          "status": {
            "type": "string",
            "enum": [
              "applied",
              "notApplied"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sourceId",
          "materialType",
          "status",
          "reason"
        ],
        "title": "MaterialReviewDTO"
      },
...
      "MaterialType": {
        "type": "string",
        "enum": [
          "company",
          "jobPosting"
        ],
        "title": "MaterialType"
      },
      "MaterialVersion": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "materialType": {
            "type": "string",
            "enum": [
              "company",
              "jobPosting"
            ],
            "title": "Materialtype"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "title": "Content"
          },
          "source": {
            "type": "string",
            "maxLength": 2000,
            "title": "Source"
          },
          "original": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileReference"
              },
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Original"
          },
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "materialId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Materialid"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "legacyUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacyurl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "materialType",
          "content",
          "source",
          "id",
          "materialId",
          "version",
          "createdAt"
        ],
        "title": "MaterialVersion"
      },
...
      "ModuleErrorDTO": {
        "properties": {
          "moduleKey": {
            "type": "string",
            "title": "Modulekey"
          },
          "inputSourceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputsourceid"
          },
          "errorCode": {
            "type": "string",
            "title": "Errorcode"
          },
          "userMessage": {
            "type": "string",
            "title": "Usermessage"
          },
          "canRetry": {
            "type": "boolean",
            "title": "Canretry"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "moduleKey",
          "inputSourceId",
          "errorCode",
          "userMessage",
          "canRetry"
        ],
        "title": "ModuleErrorDTO"
      },
...
      "PageDocument": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Pages"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pages"
        ],
        "title": "PageDocument"
      },
...
      "ResumeVersion": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "contextId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Contextid"
          },
          "version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Version"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "inputType": {
            "type": "string",
            "enum": [
              "file",
              "text"
            ],
            "title": "Inputtype"
          },
          "displayName": {
            "type": "string",
            "maxLength": 500,
            "title": "Displayname"
          },
          "original": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileReference"
              },
              {
                "type": "string",
                "maxLength": 100000
              },
              {
                "type": "null"
              }
            ],
            "title": "Original"
          },
          "status": {
            "type": "string",
            "enum": [
              "extracting",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "documentKind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "page",
                  "flow"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentkind"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageDocument"
              },
              {
                "$ref": "#/components/schemas/FlowDocument"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "contextId",
          "version",
          "createdAt",
          "inputType",
          "displayName",
          "status"
        ],
        "title": "ResumeVersion"
      },
...
      "Review": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "contextId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Contextid"
          },
          "resumeVersionId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumeversionid"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "input": {
            "$ref": "#/components/schemas/ContextRefs"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleErrorDTO"
            },
            "type": "array",
            "title": "Errors"
          },
          "materialReviews": {
            "items": {
              "$ref": "#/components/schemas/MaterialReviewDTO"
            },
            "type": "array",
            "title": "Materialreviews"
          },
          "resolutionChecks": {
            "items": {
              "$ref": "#/components/schemas/ResolutionCheck"
            },
            "type": "array",
            "title": "Resolutionchecks"
          },
          "resultIds": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "title": "Resultids"
          },
          "suggestionIds": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "title": "Suggestionids"
          },
          "experienceRecommendations": {
            "items": {
              "$ref": "#/components/schemas/ExperienceRecommendation"
            },
            "type": "array",
            "title": "Experiencerecommendations"
          },
          "modules": {
            "$ref": "#/components/schemas/ReviewModules"
          },
          "finalErrors": {
            "items": {
              "$ref": "#/components/schemas/ModuleErrorDTO"
            },
            "type": "array",
            "title": "Finalerrors"
          },
          "finalFailed": {
            "type": "boolean",
            "title": "Finalfailed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "contextId",
          "resumeVersionId",
          "createdAt",
          "input",
          "status",
          "errors",
          "materialReviews",
          "resolutionChecks",
          "resultIds",
          "suggestionIds",
          "modules",
          "finalErrors",
          "finalFailed"
        ],
        "title": "Review"
      },
...
      "ReviewMaterial": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 241,
            "minLength": 1,
            "title": "Id"
          },
          "reviewId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewid"
          },
          "materialVersionId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Materialversionid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "reviewId",
          "materialVersionId"
        ],
        "title": "ReviewMaterial"
      },
...
      "SavedExperienceDocument": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "contextId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Contextid"
          },
          "experienceId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Experienceid"
          },
          "input": {
            "$ref": "#/components/schemas/app__workspace__models__WritingInput"
          },
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "minLength": 1,
            "title": "Markdown"
          },
          "summary": {
            "type": "string",
            "maxLength": 10000,
            "title": "Summary"
          },
          "questions": {
            "items": {
              "type": "string",
              "maxLength": 4000
            },
            "type": "array",
            "maxItems": 20,
            "title": "Questions"
          },
          "sourceNotes": {
            "items": {
              "$ref": "#/components/schemas/SourceNote-Input"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Sourcenotes"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "contextId",
          "experienceId",
          "input",
          "markdown",
          "summary",
          "questions",
          "sourceNotes",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "SavedExperienceDocument"
      },
...
      "SuggestionEntry": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "contextId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Contextid"
          },
          "reviewId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewid"
          },
          "resumeVersionId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumeversionid"
          },
          "content": {
            "$ref": "#/components/schemas/SuggestionDTO"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "up",
                  "down"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "decision": {
            "type": "string",
            "enum": [
              "open",
              "resolve",
              "skip"
            ],
            "title": "Decision"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionCheck"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "contextId",
          "reviewId",
          "resumeVersionId",
          "content",
          "rating",
          "decision"
        ],
        "title": "SuggestionEntry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}